pyproject.toml
[project]
name = "order-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets==11.0.3",
]
